rust: Add rust interface for parsing JSON
For plugin commands the command definitions are stored in a JSON file as
part of the plugin. Instead of doing even more string parsing in C the
Rust library provides an interface to the popular Serde library.
The Rust interface provides a few FFI-safe definitions to pass data back
and forth but does all the parsing itself.

The library is built statically, but some extra work is done in build.rs
to help it link into the existing Autoconf more easily. This is heavily
based on the libtool crate[0] with some improvements.

[0] https://docs.rs/libtool/0.1.1/libtool/

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>