ui/ncurses: Present plugin commands in nc-plugin
Update the Plugin screen to present commands and their arguments if
available. The user may modify the arguments and run the commands as
displayed in the same way an executable may have been run previously.
An extra field is also added to provide arbitrary arguments to a given
plugin executable.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>